Diverse Modifikationen fuer das Hinzufuegen und Entfernen von Standard-Mitgliedern vorgenommen. Eine identische Person kann auch mehrmals im Studio vorkommen und erhält dann trotzdem jedes mal eine neue Mitgliedsnummer. Das Entfernen eines Mitglieds erfolgt NUR noch über die Mitgliedsnummer. Hinweis an den Benutzer, wenn eine fehlerhafte Benutzereingabe stattgefunden hat, bspw. die Mitglieds- nummer nicht (mehr) existent ist. Beispielhaft kann das in der Main- Methode bzw. wenn man dort die richtige Untermethode aufruft getestet werden.

+++ src/de/marcelhuber/fitnessstudio/Trainer.java
@@ -49,7 +49,7 @@ class Trainer extends Person {
         spezialisierungsBereiche.add(neuerSpezialisierungsBereich);
     }
 
-    long getAnzahlDerSpezialGebiete() {
+    int getAnzahlDerSpezialGebiete() {
         return spezialisierungsBereiche.size();
     }
